try catch for service, validate for email, phone

+++ Areas/Order/Pages/ConfirmPackage.cshtml.cs
@@ -114,12 +114,16 @@ public class ConfirmPackageModel : PageModel
 
     public IActionResult OnPostUpdateStatusOrder(int? orderId, bool? isFromCart)
     {
-        if (isFromCart==true)
+        if(Order.ShipFee> 0)
         {
-            cartService.DeleteBookOfOrder((int)orderId, (int)HttpContext.Session.GetInt32("Id"));
+            if (isFromCart == true)
+            {
+                cartService.DeleteBookOfOrder((int)orderId, (int)HttpContext.Session.GetInt32("Id"));
+            }
+            orderService.UpdateStatusOrder(orderService.GetOrdersByAccountId((int)HttpContext.Session.GetInt32("Id"), StatusId).FirstOrDefault().Id, 1);
+            return RedirectToPage("/OrderListFillAll");
         }
-        orderService.UpdateStatusOrder(orderService.GetOrdersByAccountId((int)HttpContext.Session.GetInt32("Id"), StatusId).FirstOrDefault().Id, 1);
-        return RedirectToPage("/OrderListFillAll");
+        return RedirectToPage("./ConfirmPackage");
     }
     public IActionResult OnPostChangeInformation()
     {
